Fix a store-queue related issue in the primitive submission code.
The primitive list submission code in libgl never sets the QACR0 or QACR1
registers when it uses the store queues. Thus, if any other code accesses the
store queues, rendering may well grind to a halt.

This fixes the issue by setting QACR0 and QACR1 each frame.

Note: This solution is still not thread-safe. Of course, many GL implementations
are not thread-safe at all...